+++ spec/support/factories.rb
@@ -7,6 +7,10 @@ FactoryGirl.define do
     sequence(:name) { |n| "category #{n}" }
   end
 
+  factory :default_without_depth do
+    sequence(:name) { |n| "category #{n}" }
+  end
+
   factory :scoped do
     sequence(:scope_type) { |n| "type_#{n}" }
     sequence(:name) { |n| "category #{n}" }